+++ projects/ng-virtual-list/src/lib/const/index.ts
@@ -26,6 +26,8 @@ export const NAVIGATION_BY_KEYBOARD_TIMER = 50;
 
 export const DEFAULT_SNAP_TO_ITEM = false;
 
+export const DEFAULT_SCROLLING_ONE_BY_ONE = false;
+
 export const DEFAULT_MOTION_BLUR = 0.25;
 
 export const DEFAULT_MAX_MOTION_BLUR = 20;