+++ projects/ng-virtual-list/CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 - Added snapToItem property.
 - Added snapToItemAlign property.
 - Added snapToItem parameter to animationParams property.
+- Added scrollingOneByOne property.
 
 ## [20.11.10] - 2026-04-13